Make Sample Work
