+++ dart-sample/app/src/main/AndroidManifest.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    android:versionCode="1"
+    android:versionName="1.0"
+    package="com.f2prateek.dart.example">
+
+  <uses-sdk android:minSdkVersion="11" android:targetSdkVersion="19"/>
+
+  <application
+      android:label="@string/app_name"
+      android:name=".SampleApp">
+
+    <activity
+        android:label="@string/app_name"
+        android:name=".MainActivity">
+      <intent-filter>
+        <action android:name="android.intent.action.MAIN"/>
+        <category android:name="android.intent.category.LAUNCHER"/>
+        <category android:name="android.intent.category.DEFAULT"/>
+      </intent-filter>
+    </activity>
+    <activity android:name=".SampleActivity"
+        android:label="@string/sampleActivityTitle">
+    </activity>
+  </application>
+</manifest>
