+++ dart-sample/app/src/main/res/values/strings.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+  ~ Copyright 2013 Jake Wharton
+  ~ Copyright 2014 Prateek Srivastava (@f2prateek)
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~        http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<resources>
+  <string name="app_name">Dart</string>
+  <string name="launchSampleActivity">Launch Sample Activity</string>
+  <string name="launchNavigationActivity">Launch Navigation activity</string>
+  <string name="launchNavigationService">Launch Navigation Service</string>
+  <string name="sampleActivityTitle">Sample Activity</string>
+</resources>
