+++ dart-sample/module1/build.gradle
@@ -1,0 +1,40 @@
+apply plugin: 'com.android.library'
+apply plugin: 'com.jakewharton.butterknife'
+
+buildscript {
+    repositories {
+        jcenter()
+        google()
+        mavenCentral()
+    }
+
+    dependencies {
+        classpath deps.butterknife.gradlePlugin
+    }
+}
+
+android {
+    compileSdkVersion versions.compileSdk
+
+    defaultConfig {
+        minSdkVersion versions.minSdk
+    }
+}
+
+dependencies {
+    api deps.android.appCompat
+    api deps.dart.runtime
+    api deps.henson.runtime
+    api deps.parceler.runtime
+    api deps.butterknife.runtime
+    api deps.butterknife.annotations
+
+    annotationProcessor deps.parceler.compiler
+    annotationProcessor deps.dart.compiler
+    annotationProcessor deps.henson.compiler
+    annotationProcessor deps.butterknife.compiler
+
+    testImplementation deps.junit
+    testImplementation deps.robolectric
+    testImplementation deps.fest
+}
