+++ dart-sample/module1/src/main/res/layout/activity_navigation.xml
@@ -1,0 +1,88 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+  ~ Copyright 2013 Jake Wharton
+  ~ Copyright 2014 Prateek Srivastava (@f2prateek)
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~        http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:padding="8dp">
+    <TextView
+      android:id="@+id/default_key_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/string_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/int_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/parcelable_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/optional_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/parcel_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/list_parcel_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/sparse_array_parcel_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+  <TextView
+      android:id="@+id/default_extra"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center"
+      style="@android:style/TextAppearance.Large"
+      />
+</LinearLayout>
