--- dart-sample/src/main/java/com/f2prateek/dart/example/ExampleParcel.java
@@ -1,19 +1,0 @@
-package com.f2prateek.dart.example;
-
-import org.parceler.Parcel;
-import org.parceler.ParcelConstructor;
-
-@Parcel
-public class ExampleParcel {
-
-  String name;
-
-  @ParcelConstructor
-  public ExampleParcel(String name) {
-    this.name = name;
-  }
-
-  public String getName() {
-    return name;
-  }
-}
