--- dart-sample/src/main/res/layout/activity_main.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-  ~ Copyright 2013 Jake Wharton
-  ~ Copyright 2014 Prateek Srivastava (@f2prateek)
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-  <Button
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:text="@string/launch"
-      android:id="@+id/button"/>
-
-  <Button
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:text="@string/launch2"
-      android:id="@+id/button2"/>
-
-</LinearLayout>
